PIM-3901: standardize the use of option resolvers in setters and copiers
