+++ src/Pim/Bundle/CatalogBundle/Updater/Copier/SimpleSelectAttributeCopier.php
@@ -43,11 +43,11 @@ class SimpleSelectAttributeCopier extends AbstractAttributeCopier
         AttributeInterface $toAttribute,
         array $options = []
     ) {
-        // TODO: option resolver for this
+        $options = $this->resolver->resolve($options);
         $fromLocale = $options['from_locale'];
-        $toLocale = $options['from_locale'];
+        $toLocale = $options['to_locale'];
         $fromScope = $options['from_scope'];
-        $toScope = $options['from_scope'];
+        $toScope = $options['to_scope'];
 
         $this->checkLocaleAndScope($fromAttribute, $fromLocale, $fromScope, 'base');
         $this->checkLocaleAndScope($toAttribute, $toLocale, $toScope, 'base');
