+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/BooleanAttributeSetter.php
@@ -40,7 +40,7 @@ class BooleanAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'boolean');
         $this->checkData($attribute, $data);
