+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/DateAttributeSetter.php
@@ -42,7 +42,7 @@ class DateAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'date');
         $data = $this->formatData($attribute, $data);
