+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/MediaAttributeSetter.php
@@ -68,7 +68,7 @@ class MediaAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'media');
         $this->checkData($attribute, $data);
