+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/MetricAttributeSetter.php
@@ -60,7 +60,7 @@ class MetricAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'metric');
         $this->checkData($attribute, $data);
