+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/NumberAttributeSetter.php
@@ -43,7 +43,7 @@ class NumberAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'number');
         $this->checkData($attribute, $data);
