+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/PriceCollectionAttributeSetter.php
@@ -59,7 +59,7 @@ class PriceCollectionAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'prices collection');
         $this->checkData($attribute, $data);
