+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/TextAttributeSetter.php
@@ -42,7 +42,7 @@ class TextAttributeSetter extends AbstractAttributeSetter
         $data,
         array $options = []
     ) {
-        $this->resolver->resolve($options);
+        $options = $this->resolver->resolve($options);
         $this->checkLocaleAndScope($attribute, $options['locale'], $options['scope'], 'text');
         $this->checkData($attribute, $data);
